Add location tracking grid and enemy ai
