+++ include/game/game.hpp
@@ -3,6 +3,7 @@
 #include "controller/input/input_state.hpp"
 #include "controller/persistence/config_game.hpp"
 #include "controller/persistence/persisted_game.hpp"
+#include "controller/state/state_transition_action.hpp"
 #include "game/ecs/registry.hpp"
 #include "game/ecs/systems/animation_system.hpp"
 #include "game/ecs/systems/camera_system.hpp"
