+++ src/controller/state/state.cpp
@@ -3,7 +3,9 @@
 #include "controller/input/mouse_util.hpp"
 #include "controller/persistence/persistence_manager.hpp"
 #include "view/text.hpp"
+#ifdef LOG_STDOUT
 #include <iostream>
+#endif
 
 namespace controller {
 
@@ -255,8 +257,9 @@ std::unique_ptr<GameplayState> GameplayState::createLoadedGameplay()
     if (persistedGame.has_value()) {
         return std::unique_ptr<GameplayState>(new GameplayState(persistedGame.value()));
     }
-
+#ifdef LOG_STDOUT
     std::cout << "No saved game found, starting new game instead." << std::endl;
+#endif
     return createNewGameplay();
 }
