+++ src/game/systems/enemy_ai.cpp
@@ -2,9 +2,10 @@
 #include "game/ecs/components/enemy_tag.hpp"
 #include "game/ecs/components/player_tag.hpp"
 #include "game/ecs/components/position.hpp"
-#include "game/ecs/components/sprite.hpp"
+#include "game/ecs/components/stats.hpp"
 #include "game/ecs/components/velocity.hpp"
 #include "game/location_table.hpp"
+#include "view/sprite.hpp"
 #include <cmath>
 #include <iostream>
 
@@ -18,10 +19,11 @@ void EnemyAI::update(Registry &registry, LocationTable &locationTable)
         const Position playerPos = registry.getComponent<Position>(player);
         const Vec2 posP{playerPos.x, playerPos.y};
 
-        for (auto enemy : registry.view<Velocity, Sprite, Position, EnemyTag>()) {
-            EnemyTag &enemyTag = registry.getComponent<EnemyTag>(enemy);
+        for (auto enemy : registry.view<Velocity, view::Sprite, Position, EnemyTag, EnemyStats>()) {
+            // EnemyTag &enemyTag = registry.getComponent<EnemyTag>(enemy);
             Velocity &velocity = registry.getComponent<Velocity>(enemy);
             Position &enemyPos = registry.getComponent<Position>(enemy);
+            EnemyStats &enemyStats = registry.getComponent<EnemyStats>(enemy);
             Vec2 posE{enemyPos.x, enemyPos.y};
 
             // Set movement direction exactly towards player
@@ -42,7 +44,7 @@ void EnemyAI::update(Registry &registry, LocationTable &locationTable)
             v.normalize();
             v += repelOffset;
 
-            v.setLenght(enemyTag.moveSpeed);
+            v.setLenght(enemyStats.moveSpeed);
             velocity.dx = v.x; // TODO: into Vec2
             velocity.dy = v.y; // TODO: into Vec2
         }
