+++ src/game/systems/movement_system.cpp
@@ -1,9 +1,9 @@
 #include "game/ecs/systems/movement_system.hpp"
 #include "game/ecs/components/position.hpp"
 #include "game/ecs/components/velocity.hpp"
-
+#ifdef LOG_STDOUT
 #include <iostream>
-
+#endif
 namespace game {
 
 void MovementSystem::update(Registry &registry, float dt)
@@ -17,9 +17,11 @@ void MovementSystem::update(Registry &registry, float dt)
         position.x += velocity.dx * dt;
         position.y += velocity.dy * dt;
 
+#ifdef LOG_STDOUT
         if (position.x != before.x || position.y != before.y) {
-            // std::cout << "Entity " << entity << " moved to (" << position.x << ", " << position.y << ")\n";
+            std::cout << "Entity " << entity << " moved to (" << position.x << ", " << position.y << ")\n";
         }
+#endif
     }
 }
